+++ packages/sdk-core/CHANGELOG.md
@@ -5,12 +5,17 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [0.4.5] - 2022-07-13
+
+## Added
+- Support for SetEvent and Subgraph v1.4.4
+
 ### Changed
 - `SFError` refactor to be more conventional. It inherits `Error` and uses `cause` to wrap internal errors.
 - Remove serialized internal error from the messages of `SFError` (it's now just included in the `.cause` property)
 
 ### Breaking
-- `SFError.errorObject` renamed to `SFError.cause`  
+- `SFError.errorObject` renamed to `SFError.cause`
 
 ## [0.4.4] - 2022-06-30